backend/requirements.txt
# AlphaZero Chess v1.0.0 - Python Backend Dependencies
# =====================================================
# This file contains all required Python packages for the AlphaZero Chess backend.
# Install with: pip install -r requirements.txt
#
# Minimum Requirements:
# - Python 3.10+
# - pip 22+
#
# For GPU support (optional):
# - CUDA Toolkit 11.8+ (NVIDIA GPUs only)
# - Install PyTorch with CUDA: pip install torch --index-url https://download.pytorch.org/whl/cu118
# =====================================================

# Core Framework
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
python-dotenv==1.1.1
python-multipart==0.0.20

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
PyJWT==2.10.1
cryptography==46.0.1
ecdsa==0.19.1

# Data Validation & Serialization
pydantic==2.11.9
pydantic_core==2.33.2
email-validator==2.3.0

# AI/ML Core
torch
numpy==2.3.3
pandas==2.3.3

# Chess Engine
chess==1.11.2
python-chess==1.999

# HTTP Clients & Async
aiohttp==3.13.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
requests-oauthlib==2.0.0

# LLM Integration
emergentintegrations==0.1.0
openai==1.99.9
litellm==1.77.7

# Google AI
google-generativeai==0.8.5
google-genai==1.41.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2

# AWS/Cloud
boto3==1.40.41
botocore==1.40.41
s3transfer==0.14.0
stripe==13.0.1

# CLI & Terminal
typer==0.19.2
rich==14.1.0
click==8.3.0
shellingham==1.5.4

# Utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
tenacity==9.1.2
tqdm==4.67.1
cachetools==6.2.0
filelock==3.19.1
fsspec==2025.9.0
pillow==11.3.0

# Tokenization
tiktoken==0.12.0
tokenizers==0.22.1
pytokens==0.1.10

# Data Processing
PyYAML==6.0.3
Jinja2==3.1.6
MarkupSafe==3.0.3
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2

# JSON Schema & Validation
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.36.2
rpds-py==0.27.1

# Networking & Protocols
h11==0.16.0
httplib2==0.31.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
frozenlist==1.8.0
multidict==6.7.0
yarl==1.22.0
propcache==0.4.0

# Cryptography & Encoding
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
pycparser==2.23
pyasn1==0.6.1
pyasn1_modules==0.4.2
rsa==4.9.1

# Protocol Buffers & gRPC
protobuf==5.29.2
proto-plus==1.26.1

# OAuth & Authentication
oauthlib==3.3.1

# Development & Testing
pytest==8.4.2
pluggy==1.6.0
iniconfig==2.1.0
packaging==25.0

# Code Quality & Linting
flake8==7.3.0
black==25.9.0
mypy==1.18.2
mypy_extensions==1.1.0
isort==6.0.1
pycodestyle==2.14.0
pyflakes==3.4.0
mccabe==0.7.0

# Type Checking & Annotations
typing_extensions==4.15.0
typing-inspection==0.4.1
annotated-types==0.7.0

# System & Process Management
psutil==7.1.0
watchfiles==1.1.0
distro==1.9.0

# Hugging Face
huggingface-hub==0.35.3
hf-xet==1.1.10

# Miscellaneous
attrs==25.4.0
anyio==4.11.0
sniffio==1.3.1
idna==3.10
six==1.17.0
jmespath==1.0.1
pyparsing==3.2.5
regex==2025.9.18
pathspec==0.12.1
platformdirs==4.4.0
zipp==3.23.0
importlib_metadata==8.7.0
jiter==0.11.0
fastuuid==0.13.5
jq==1.10.0
s5cmd==0.2.0

# Note: torch is listed without version to allow flexible installation
# For CPU only: pip install torch
# For CUDA 11.8: pip install torch --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch --index-url https://download.pytorch.org/whl/cu121
